requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
gunicorn==20.1.0
psycopg2-binary==2.9.6
